up to 4.23

+++ part4/backend/controllers/blogs.js
@@ -2,6 +2,8 @@
 const blogRouter = require('express').Router()
 const Blog = require('../models/blog')
 const User = require('../models/user')
+const jwt = require('jsonwebtoken')
+
 
 blogRouter.get('/', async (request, response) => {
     const blogs = await Blog
@@ -14,7 +16,13 @@ blogRouter.get('/', async (request, response) => {
 blogRouter.post('/', async (request, response) => {
   const body = request.body
 
-  const user = await User.findById(body.userId)
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)  
+  if (!decodedToken.id) {    
+    return response.status(401).json({ error: 'token invalid' })  
+  }  
+  // const user = await User.findById(decodedToken.id)
+
+  const user = request.user
 
   const blog = new Blog({
     title: body.title,
@@ -33,7 +41,17 @@ blogRouter.post('/', async (request, response) => {
 })
 
 blogRouter.delete('/:id', async (request, response) => {
+  const blog = await Blog.findById(request.params.id)
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)  
+
+  console.log(request.user.id)
+
+  if (decodedToken.id !== blog.user.toString()) {  
+    return response.status(401).json({ error: 'invalid token for deleting' })  
+  }  
+
   await Blog.findByIdAndDelete(request.params.id)
+
   response.status(204).end()
 })
